# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Property Management System
  version: 1.0.0
  description: |
    Source of truth for the boundary between the Angular application and the Go service.
    The Go server interface and the Angular client are both generated from this document;
    neither side hand-writes these types (Constitution III).

    Authorization is decided on the server for every operation. `x-required-role` states
    the requirement; an operation marked `admin` is refused for a manager regardless of
    how the request is made (Constitution I). Configuration — property types, areas,
    custom field definitions, and changing a property's reference code — are `admin`.
    Everything else is `manager`, which an administrator also satisfies.

    Sensitive custom fields never appear in this document as a filterable field. The
    server refuses a filter naming one even if a client constructs the request by hand
    (FR-027s4, Constitution VII).

    Every response body uses Western Arabic digits. Messages shown to users are Arabic.
    All timestamps are RFC 3339 in UTC; the client renders them in Africa/Cairo.
servers:
  - url: https://localhost:8443/api/v1
    description: Local development. TLS only — there is no plaintext listener.
...
security:
  - sessionCookie: []
paths:
...
  /properties/search:
    post:
      tags: [search]
      operationId: searchProperties
      summary: Advanced search across built-in and custom fields
      description: |
        Filters combine with AND; an unset filter constrains nothing.

        A filter naming a sensitive field is refused with 400, not silently dropped
        (Constitution VII). The server resolves every `fieldId` against the definitions
        before building the query; an unknown id is refused.
      x-required-role: manager
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdvancedSearch' }
      responses:
        '200':
          description: A page of matching properties.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PropertyPage' }
        '400':
          description: |
            An unknown field, a sensitive field, or an operator that does not suit the
            field's type.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pms_session
      description: |
        Set by /auth/login. HttpOnly, Secure, SameSite=Strict. Not readable by
        JavaScript and never placed in a URL. Carries identity only; role and active
        state are read from stored data on every request.
...
  schemas:
...
    PropertyCode:
      type: string
      minLength: 1
      maxLength: 32
      description: |
        Unique across the register, active and archived alike, compared after
        trimming, case folding, and Arabic normalisation. Never reused.
    PropertyName:
      type: string
      minLength: 2
      maxLength: 120
      description: |
        Descriptive, not identifying. Two properties may share a name.
    Lookup:
      type: object
      required: [id, label]
      properties:
        id:    { type: string, format: uuid }
        label: { type: string, minLength: 1, maxLength: 60 }
...
    PropertySummary:
      type: object
      required: [id, code, name, propertyType, area, isArchived]
      description: One row of the list. Custom values are not included here.
      properties:
        id:           { type: string, format: uuid }
        code:         { $ref: '#/components/schemas/PropertyCode' }
        name:         { $ref: '#/components/schemas/PropertyName' }
        propertyType: { $ref: '#/components/schemas/Lookup' }
        area:         { $ref: '#/components/schemas/Lookup' }
        isArchived:   { type: boolean }
...
    PropertyPage:
      type: object
      required: [items, page, pageSize, totalItems]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/PropertySummary' }
        page:       { type: integer }
        pageSize:   { type: integer }
        totalItems: { type: integer }
    SearchOperator:
      type: string
      enum: [contains, equals, includesAll, isTrue, isFalse]
      description: |
        `contains` suits text; `equals` a dropdown; `includesAll` a multiselect;
        `isTrue`/`isFalse` a checkbox. The server checks the pairing and refuses a
        mismatch with 400.
    CustomFieldFilter:
      type: object
      required: [fieldId, operator]
      properties:
        fieldId:   { type: string, format: uuid }
        operator:  { $ref: '#/components/schemas/SearchOperator' }
        text:      { type: [string, 'null'] }
        choiceId:  { type: [string, 'null'], format: uuid }
        choiceIds:
          type: array
          items: { type: string, format: uuid }
    AdvancedSearch:
      type: object
      properties:
        q:              { type: string, maxLength: 120 }
        propertyTypeId: { type: string, format: uuid }
        areaId:         { type: string, format: uuid }
        includeArchived: { type: boolean, default: false }
        page:     { type: integer, minimum: 1, default: 1 }
        pageSize: { type: integer, enum: [10, 25, 50, 100], default: 25 }
        customFilters:
          type: array
          items: { $ref: '#/components/schemas/CustomFieldFilter' }
    Error:
      type: object
      required: [code, message]
      description: |
        The single error shape used by every operation in this document
        (Constitution III). `message` is Arabic and states what the user should do
        next. It never carries internal failure detail, a key, a wrapped key, or a
        decrypted value (Constitution VII).
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - invalid_credentials
            - too_soon
            - not_authenticated
            - forbidden
            - not_found
            - conflict
            - in_use
            - version_conflict
            - archived
            - password_change_required
            - internal_error
        message:
          type: string
          description: Arabic, user-facing.
        fields:
          type: [object, 'null']
          additionalProperties:
            type: string
          description: Field-level Arabic messages, keyed by field name, for form errors.
        retryAfterSeconds:
          type: [integer, 'null']
          minimum: 1
          maximum: 60
          description: Seconds remaining before another attempt is accepted (sign-in throttling).
  responses:
    Error:
      description: The request was refused. Body is the single error shape.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'